server/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database - PostgreSQL
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Vector Database - ChromaDB
chromadb==0.4.22
sentence-transformers==2.3.1

# Machine Learning
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
joblib==1.3.2

# Google Gemini LLM
google-generativeai==0.3.2

# Security - JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging & Monitoring
python-json-logger==2.0.7
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For testing async endpoints

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# CORS

